requirements.txt
tensorflow
flask
pillow
pandas
matplotlib
numpy
requests